+++ src/modules/scenarioeditor/ToolPanel.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
+import { computed, Ref, ref, watch } from "vue";
 import {
+  IconChevronRight,
   IconCursorDefaultOutline,
   IconCursorMove,
   IconDotsHorizontal,
@@ -8,21 +10,24 @@ import {
   IconMapMarkerPath,
   IconVectorPolylineEdit,
 } from "@iconify-prerendered/vue-mdi";
+import { Float } from "@headlessui-float/vue";
+import OLMap from "ol/Map";
+import { Popover, PopoverButton, PopoverPanel } from "@headlessui/vue";
 import FloatingPanel from "@/components/FloatingPanel.vue";
 import PanelSection from "@/components/PanelSection.vue";
 import MilSymbol from "@/components/MilSymbol.vue";
 import { SymbolItem, SymbolValue } from "@/types/constants";
-import { computed, ref } from "vue";
 import { useGetMapLocation } from "@/composables/geoMapLocation";
-import OLMap from "ol/Map";
 import { useGeoStore, useUnitSettingsStore } from "@/stores/geoStore";
 import { injectStrict } from "@/utils";
 import { activeScenarioKey, activeUnitKey, sidcModalKey } from "@/components/injects";
 import { useToggle } from "@vueuse/core";
 import PanelToggle from "@/components/PanelToggle.vue";
-import { Sidc } from "@/symbology/sidc";
+import { SID_INDEX, Sidc } from "@/symbology/sidc";
 import PanelButton from "@/components/PanelButton.vue";
 import PanelSymbolButton from "@/components/PanelSymbolButton.vue";
+import { echelonItems } from "@/symbology/helpers";
+import { SID, SidValue } from "@/symbology/values";
 
 const [addMultiple, toggleAddMultiple] = useToggle(false);
 
@@ -36,10 +41,7 @@ const {
 const { getModalSidc } = injectStrict(sidcModalKey);
 
 const activeUnit = computed(
-  () =>
-    (activeUnitId.value &&
-      unitActions.expandUnit(state.getUnitById(activeUnitId.value))) ||
-    null
+  () => (activeUnitId.value && state.getUnitById(activeUnitId.value)) || null
 );
 
 const geoStore = useGeoStore();
@@ -54,27 +56,35 @@ const icons: SymbolValue[] = [
   { code: "160600", text: "Combat Service Support" },
 ];
 
-const customIcon = ref<SymbolValue>({ code: "10031000141211004600", text: "Infantry" });
+const customIcon = ref<SymbolValue>({ code: "10031000141211000000", text: "Infantry" });
+
+const currentSid = ref<string>(SID.Friend);
+const currentEchelon = ref<string>("16");
 
 const iconItems = computed(() => {
-  const sid = activeUnit.value?.sidc ? activeUnit.value?.sidc[3] : "3";
-
   return icons.map(({ code, text }): SymbolItem => {
     return {
       code,
       text,
-      sidc: "100" + sid + "10" + "00" + "00" + code + "0000",
+      sidc: "100" + currentSid.value + "10" + "00" + "00" + code + "0000",
     };
   });
 });
+const echelons = computed(() => echelonItems(currentSid.value as SidValue));
 
 const activeSidc = ref<string | null>(null);
-const activeEchelon = ref<string | null>(null);
+const echelonSidc = computed(
+  () => "100" + currentSid.value + "10" + "00" + currentEchelon.value + "0000000000"
+);
+
+watch(activeUnit, (unit) => {
+  if (!unit) return;
+  currentSid.value = unit.sidc[SID_INDEX];
+});
 
 const customSidc = computed(() => {
-  const sid = activeUnit.value?.sidc ? activeUnit.value?.sidc[3] : "3";
   const parsedSidc = new Sidc(customIcon.value.code);
-  parsedSidc.standardIdentity = sid;
+  parsedSidc.standardIdentity = currentSid.value;
   parsedSidc.emt = "00";
   parsedSidc.hqtfd = "0";
   return parsedSidc.toString();
@@ -101,8 +111,10 @@ onGetLocation((location) => {
   groupUpdate(() => {
     if (!activeUnit.value) return;
     const name = `${(activeUnit.value?.subUnits?.length ?? 0) + 1}`;
+    const sidc = new Sidc(activeSidc.value!);
+    sidc.emt = currentEchelon.value;
     const unitId = unitActions.createSubordinateUnit(activeUnit.value.id, {
-      sidc: activeSidc.value!,
+      sidc: sidc.toString(),
       name,
     });
     unitId && addUnitPosition(unitId, location);
@@ -123,6 +135,11 @@ async function handleChangeSymbol() {
     customIcon.value.code = newSidcValue;
     addUnit(customSidc.value);
   }
+}
+
+function selectEchelon(sidc: string, closePopover: (ref?: Ref | HTMLElement) => void) {
+  currentEchelon.value = new Sidc(sidc).emt;
+  closePopover();
 }
 </script>
 <template>
@@ -160,7 +177,7 @@ async function handleChangeSymbol() {
         </button>
       </PanelSection>
       <PanelSection label="Add unit">
-        <div class="mt-1 grid grid-cols-2 gap-2">
+        <div class="mt-1 grid grid-cols-2 justify-items-center gap-2.5">
           <button
             type="button"
             class="flex items-center justify-center hover:drop-shadow hover:sepia disabled:opacity-50 disabled:hover:sepia-0"
@@ -182,14 +199,36 @@ async function handleChangeSymbol() {
             @click="addUnit(sidc)"
             :disabled="!activeUnitId"
           />
+          <MilSymbol :size="24" :sidc="echelonSidc || ''" class="self-center" />
+          <Popover as="template">
+            <Float placement="right-start" :offset="12" flip shift portal>
+              <PopoverButton as="template">
+                <PanelButton title="Select echelon" class=""
+                  ><IconChevronRight class="h-5 w-5"
+                /></PanelButton>
+              </PopoverButton>
+              <PopoverPanel focus v-slot="{ close }">
+                <FloatingPanel class="grid grid-cols-4 justify-items-center gap-2 p-2">
+                  <PanelSymbolButton
+                    class="self-end"
+                    v-for="{ sidc, text } in echelons"
+                    :key="sidc"
+                    :sidc="sidc"
+                    :title="text"
+                    @click="selectEchelon(sidc, close)"
+                  />
+                </FloatingPanel>
+              </PopoverPanel>
+            </Float>
+          </Popover>
         </div>
       </PanelSection>
       <PanelSection label="Path" class="">
+        {{ currentSid }}
         <div class="grid grid-cols-2 gap-2">
           <PanelToggle v-model="unitSettings.showHistory" title="Show unit path">
             <IconMapMarkerPath class="h-5 w-5" aria-hidden="true" />
           </PanelToggle>
-
           <PanelToggle
             title="Edit path"
             v-model="unitSettings.editHistory"